keypb: Remove the Secret wrapper message.
It is not safe to rely on the protobuf decoder to decode a message that may
have been mangled by decryption. In particular, the decoder can fairly easily
be fooled into accepting a semantically invalid message, e.g., by ignoring
unknown field tags, which is WAI for protobuf but dangerous in this context.

Instead, do something simpler: Pack the checksum and the payload into a flat
array and encrypt them that way. Also while I'm here, switch to using CRC32.
Remove the Secret message type and update documentation.